Add virtual method ChSparseMatrix::Compress.
- this is an optional optimization step that a derived class may perform after elements have been inserted in the matrix.  If implemented, it must return true if it makes any modifications
- implement all logic related to the sparsity pattern lock inside ChCSR3Matrix::Compress().  The flag m_lock_broken is now a private state flag.
- this removes all dependencies of ChSolverMKL on a particular sparse matrix implementation

+++ src/chrono_mkl/ChSolverMKL.h
@@ -40,8 +40,8 @@ class ChApiMkl ChSolverMKL : public ChSolver {
     ChMatrixDynamic<double> rhs;
     ChMatrixDynamic<double> sol;
     ChMklEngine mkl_engine;
-    size_t n;
-    size_t nnz;
+    int n;
+    int nnz;
 
     bool sparsity_pattern_lock;
     bool use_perm;
@@ -71,7 +71,7 @@ class ChApiMkl ChSolverMKL : public ChSolver {
     void SetPreconditionedCGS(bool on_off, int L) { mkl_engine.SetPreconditionedCGS(on_off, L); }
 
     /// Set the number of non-zero entries in the problem matrix.
-    void SetMatrixNNZ(size_t nnz_input) { nnz = nnz_input; }
+    void SetMatrixNNZ(int nnz_input) { nnz = nnz_input; }
 
     /// Reset timers for internal phases in Solve and Setup.
     void ResetTimers() {
